+++ Makefile
@@ -4,6 +4,5 @@ default:
 
 clean:
 	@rm dist/*
-	@rm demos/globaldemo/minimalcomps_1.0.0.min.js
-	@rm demos/moduledemo/minimalcomps_1.0.0.min.mjs
+	@rm demos/globaldemo/minimalcomps*js
