+++ demos/globaldemo/index.html
@@ -10,7 +10,7 @@
     </style>
   </head>
   <body>
-    <script src="./minimalcomps_1.0.0.min.js"></script>
+    <script src="./minimalcomps_1.1.0.min.js"></script>
     <script src="./main.js"></script>
   </body>
 </html>
