+++ demos/moduledemo/main.js
@@ -1,12 +1,25 @@
-import { Panel, Button, Checkbox, Label, ProgressBar, RadioButton, TextInput, TextArea, HSlider, VSlider, ColorPicker, NumericStepper, Dropdown, Image, TextBox, Canvas } from "./minimalcomps.1.0.0.min.mjs";
+import {
+  Button,
+  Canvas, Checkbox, ColorPicker,
+  Dropdown,
+  HSlider,
+  Image,
+  Label,
+  NumericStepper,
+  Panel, ProgressBar,
+  RadioButton,
+  TextArea, TextBox, TextInput,
+  VBox, VSlider,
+} from "./minimalcomps_1.1.0.min.mjs";
 
 const panel = new Panel(document.body, 40, 40, 580, 400);
 
-new Button(panel, 20, 20, "Button")
+new Button(panel, 20, 20, "Button");
 new Checkbox(panel, 20, 50, "Checkbox");
-new RadioButton(panel, 20, 80, "group", "RadioButton 1", true);
-new RadioButton(panel, 20, 100, "group", "RadioButton 2", false);
-new RadioButton(panel, 20, 120, "group", "RadioButton 3", false);
+const vbox = new VBox(panel, 20, 80, 10);
+new RadioButton(vbox, 0, 0, "group", "RadioButton 1", true);
+new RadioButton(vbox, 0, 0, "group", "RadioButton 2", false);
+new RadioButton(vbox, 0, 0, "group", "RadioButton 3", false);
 new Label(panel, 20, 150, "Label");
 new ProgressBar(panel, 20, 180, 30, 100);
 new TextInput(panel, 20, 210, "TextInput");
@@ -19,7 +32,7 @@ new NumericStepper(panel, 220, 50, 20, 0, 100);
 const stuff = ["one", "two", "three", "four", "five"];
 new Dropdown(panel, 220, 80, stuff, 0);
 new Image(panel, 220, 120, "demo.jpg");
-new TextBox(panel, 220, 250, "TextBox - a multiline, fixed size text box for displaying non-editable text, including html.").html=true;
+new TextBox(panel, 220, 250, "TextBox - a multiline, fixed size text box for displaying non-editable text, including html.").html = true;
 const canvas = new Canvas(panel, 350, 20, 200, 100);
 canvas.context.lineWidth = 0.5;
 canvas.context.beginPath();
